pyproject.toml
[tool.poetry]
name = "simple-tg-bot"
version = "0.1.0"
description = "A simple bot example to show python skills"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
python-dotenv = "^1.1.1"
aiosqlite = "^0.21.0"
tenacity = "^9.1.2"
httpx = "^0.28.1"
python-telegram-bot = "^22.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
